Changed isort config and fixed import accordingly
