+++ tests/data/test_serialize_json.py
@@ -1,6 +1,5 @@
 from tests.data.test_common import _assert_tar_file_contents
-from unifair.data.json import JsonDataset
-from unifair.data.json import JsonDatasetToTarFileSerializer
+from unifair.data.json import JsonDataset, JsonDatasetToTarFileSerializer
 
 
 def test_json_dataset_to_tar_file_serializer_single_obj_type():
