+++ tests/data/test_serialize_pandas.py
@@ -1,7 +1,5 @@
-from tests.data.test_common import _assert_pandas_dataset_equals
-from tests.data.test_common import _assert_tar_file_contents
-from unifair.data.pandas import PandasDataset
-from unifair.data.pandas import PandasDatasetToTarFileSerializer
+from tests.data.test_common import (_assert_pandas_dataset_equals, _assert_tar_file_contents)
+from unifair.data.pandas import PandasDataset, PandasDatasetToTarFileSerializer
 
 
 def test_pandas_dataset_to_tar_file_serializer_single_obj_type():
