+++ unifair/steps/imports/gdc.py
@@ -2,8 +2,7 @@ import time
 
 import requests
 
-from unifair.core.data import JsonDocumentCollection
-from unifair.core.data import NoData
+from unifair.core.data import JsonDocumentCollection, NoData
 from unifair.core.workflow import WorkflowStep
