+++ unifair/unifair_prefect.py
@@ -5,8 +5,7 @@ from typing import Dict, List
 
 import pandas as pd
 import prefect
-from prefect import Flow
-from prefect import task
+from prefect import Flow, task
 from prefect.engine.results import LocalResult
 from prefect.executors import LocalExecutor
 from pydantic import BaseModel
